requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref=3.4.1=pyhd3eb1b0_0
ca-certificates=2021.10.8=h5b45459_0
certifi=2021.10.8=py38haa244fe_1
django=3.2.5=pyhd3eb1b0_0
django-cors-headers=3.11.0=pypi_0
djangorestframework=3.13.1=pyhd8ed1ab_0
krb5=1.19.2=h5b6d351_0
libpq=12.9=hb652d5d_1
openssl=1.1.1l=h8ffe710_0
pip=21.2.2=py38haa95532_0
psycopg2=2.8.6=py38hcd4344a_1
pymysql=1.0.2=pyhd8ed1ab_0
python=3.8.12=h6244533_0
python_abi=3.8=2_cp38
pytz=2021.3=pyhd3eb1b0_0
setuptools=58.0.4=py38haa95532_0
sqlite=3.37.0=h2bbff1b_0
sqlparse=0.4.1=py_0
typing_extensions=3.10.0.2=pyh06a4308_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py38haa95532_2
zlib=1.2.11=h8cc25b3_4
